Move testing utils to core
